You can't outrun your shadow... Reinstate the SSL switch
Turns out, nginx just won't start if you give it an SSL certificate
path that, ahem, doesn't exist. So you really do have to write the
whole config without SSL, request the certs, then write the whole
config with SSL again.

This is annoying, but it's how it was set up before, and it works, so
let's not look too hard at it.